Add shared history between shell sessions
All new commands will automatically write to the history, but shells
will not automatically re-load the history file due to a bug that puts
the entire history file in the history when calling
history-search-backward.

Use hr to reload history manually if needed.

Added an awk command that will rewrite the history file on shell load to
remove duplicates.